// src/models/external-api/study.ts
/**
 * ClinicalTrials.gov REST API
 * This API is made available to provide users meta data, statistics, and the most recent version of the clinical trials available on ClinicalTrials.gov.
 *
 * The version of the OpenAPI document: 2.0.0-draft
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AnnotationSection } from './annotationSection';
import { DerivedSection } from './derivedSection';
import { ResultsSection } from './resultsSection';
import { DocumentSection } from './documentSection';
import { ProtocolSection } from './protocolSection';

export interface Study {
  protocolSection?: ProtocolSection;
  resultsSection?: ResultsSection;
  annotationSection?: AnnotationSection;
  documentSection?: DocumentSection;
  derivedSection?: DerivedSection;
  hasResults?: any | null;
}